+++ src/fieldlayoutelements/BaseField.php
@@ -215,6 +215,7 @@ abstract class BaseField extends FieldLayoutElement
             'tipId' => $this->tipId(),
             'warningId' => $this->warningId(),
             'errorsId' => $this->errorsId(),
+            'statusId' => $this->statusId(),
             'fieldAttributes' => $this->containerAttributes($element, $static),
             'inputContainerAttributes' => $this->inputContainerAttributes($element, $static),
             'labelAttributes' => $this->labelAttributes($element, $static),
@@ -312,6 +313,17 @@ abstract class BaseField extends FieldLayoutElement
     }
 
     /**
+     * Returns the `id` if the field status message.
+     *
+     * @return string
+     * @since 3.7.29
+     */
+    protected function statusId(): string
+    {
+        return sprintf('%s-status', $this->id());
+    }
+
+    /**
      * Returns the `aria-describedby` attribute value that should be set on the focusable input(s).
      *
      * @param ElementInterface|null $element The element the form is being rendered for
@@ -324,6 +336,7 @@ abstract class BaseField extends FieldLayoutElement
     {
         $ids = array_filter([
             (!$static && $this->errors($element)) ? $this->errorsId() : null,
+            $this->statusClass($element, $static) ? $this->statusId() : null,
             $this->instructions($element, $static) ? $this->instructionsId() : null,
             $this->tip($element, $static) ? $this->tipId() : null,
             $this->warning($element, $static) ? $this->warningId() : null,