+++ src/helpers/Cp.php
@@ -471,17 +471,20 @@ class Cp
         $tipId = $config['tipId'] ?? "$id-tip";
         $warningId = $config['warningId'] ?? "$id-warning";
         $errorsId = $config['errorsId'] ?? "$id-errors";
+        $statusId = $config['statusId'] ?? "$id-status";
 
         $instructions = $config['instructions'] ?? null;
         $tip = $config['tip'] ?? null;
         $warning = $config['warning'] ?? null;
         $errors = $config['errors'] ?? null;
+        $status = $config['status'] ?? null;
 
         if (StringHelper::startsWith($input, 'template:')) {
             // Set a describedBy value in case the input template supports it
             if (!isset($config['describedBy'])) {
                 $descriptorIds = array_filter([
                     $errors ? $errorsId : null,
+                    $status ? $statusId : null,
                     $instructions ? $instructionsId : null,
                     $tip ? $tipId : null,
                     $warning ? $warningId : null,
@@ -495,7 +498,6 @@ class Cp
         $fieldset = $config['fieldset'] ?? false;
         $fieldId = $config['fieldId'] ?? "$id-field";
         $labelId = $config['labelId'] ?? "$id-" . ($fieldset ? 'legend' : 'label');
-        $status = $config['status'] ?? null;
         $label = $config['fieldLabel'] ?? $config['label'] ?? null;
 
         if ($label === '__blank__') {
@@ -568,6 +570,7 @@ class Cp
                 : '') .
             ($status
                 ? Html::beginTag('div', [
+                    'id' => $statusId,
                     'class' => ['status-badge', $status[0]],
                     'title' => $status[1],
                 ]) .